added onboarding process
that exists now. also staticd the backend and made an appcompatactivity
class for ensuring proper backend credentials and permissions are always
satisfied. immediate next step is the ui for clicking on clusters, but i
dont imagine that will be too difficult. after that fixing up the
profile page, then actually implementing the trending page